pyproject.toml
[project]
name = "spyx"
version = "0.1.20"
description = "Spyx: SNNs in JAX"
authors = [
    { name = "Kade Heckel", email = "kade.heckel@gmail.com" }
]
dependencies = [
    "optax",
    "jax_tqdm",
    "nir",
    "flax>=0.10.7",
    "grain",
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = { text = "MIT" }

[project.optional-dependencies]
loaders = [
    "tonic",
    "numba>=0.59.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.0",
]

[tool.uv.sources]
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/spyx"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
# Only apply linting to src and tests directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    # Project-specific directories to exclude
    "research",
    "docs",
    "scripts",
    "data",
    "specs",
    "site",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# Error (`E`) codes by default.
select = ["E4", "E7", "E9", "F", "B", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# ReStructuredText code blocks are supported.
#
# This is currently alphabetical by default, but will be changed to
# match the order of the configuration in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This defaults to the main `line-length` setting.
docstring-code-line-length = "dynamic"
